chore: Added a pdf folder for handling all pdf generation functions
